+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Add user access to the beacon statistics data with `smtc_modem_class_b_beacon_get_statistics` function.
 * Add user access to the `smtc_modem_factory_reset` function.
 * Add support for 2.4 Ghz multicast class C session frequency setting.
+* Add support over SMTC modem API to set LoRaWAN uplink dwell time restriction.
+* Add support over SMTC modem API to lock LoRaWAN uplink dwell time against network changes via `TxParamSetupReq`.
 
 ### Changed